Development Notes from Sighthound
This web application demo displays objects detected by a Sighthound video analytics pipeline.
The UI displays photos of objects found.
When a vehicle is found which includes a license plate and license plate string is resolved, the vehicle is displayed with an inset of the license plate string.
The selected vehicle or person is shown in the middle top part of the screen, while all vehicles/people found will show in a grid in the lower part of the screen.
You can change the selected vehicle/person by clicking on its photo in the lower grid.
Check the "Specify Sever" button to specify the server device running SIO software. This can be used to specify the hostname or IP address of a DNN Node device running sighthound analytics.
When connected to the server, the "Connected" status icon should turn green.
Click the "Show Results" button to show or save a list of CSV data captured by the application.